reconstructed of data with substantive package
