+++ app/src/main/java/com/google/samples/apps/sunflower/adapters/GalleryAdapter.kt
@@ -25,7 +25,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.google.samples.apps.sunflower.GalleryFragment
 import com.google.samples.apps.sunflower.adapters.GalleryAdapter.GalleryViewHolder
-import com.google.samples.apps.sunflower.data.UnsplashPhoto
+import com.google.samples.apps.sunflower.data.entity.UnsplashPhoto
 import com.google.samples.apps.sunflower.databinding.ListItemPhotoBinding
 
 /**
