+++ app/src/main/java/com/google/samples/apps/sunflower/adapters/GardenPlantingAdapter.kt
@@ -26,7 +26,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.google.samples.apps.sunflower.R
 import com.google.samples.apps.sunflower.HomeViewPagerFragmentDirections
-import com.google.samples.apps.sunflower.data.PlantAndGardenPlantings
+import com.google.samples.apps.sunflower.data.entity.PlantAndGardenPlantings
 import com.google.samples.apps.sunflower.databinding.ListItemGardenPlantingBinding
 import com.google.samples.apps.sunflower.viewmodels.PlantAndGardenPlantingsViewModel
 
@@ -77,15 +77,15 @@ class GardenPlantingAdapter :
 private class GardenPlantDiffCallback : DiffUtil.ItemCallback<PlantAndGardenPlantings>() {
 
     override fun areItemsTheSame(
-        oldItem: PlantAndGardenPlantings,
-        newItem: PlantAndGardenPlantings
+            oldItem: PlantAndGardenPlantings,
+            newItem: PlantAndGardenPlantings
     ): Boolean {
         return oldItem.plant.plantId == newItem.plant.plantId
     }
 
     override fun areContentsTheSame(
-        oldItem: PlantAndGardenPlantings,
-        newItem: PlantAndGardenPlantings
+            oldItem: PlantAndGardenPlantings,
+            newItem: PlantAndGardenPlantings
     ): Boolean {
         return oldItem.plant == newItem.plant
     }
