+++ app/src/main/java/com/google/samples/apps/sunflower/data/PlantDao.kt
@@ -21,6 +21,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import com.google.samples.apps.sunflower.data.entity.Plant
 
 /**
  * The Data Access Object for the Plant class.
