+++ app/src/main/java/com/google/samples/apps/sunflower/data/UnsplashPagingSource.kt
@@ -18,6 +18,7 @@ package com.google.samples.apps.sunflower.data
 
 import androidx.paging.PagingSource
 import com.google.samples.apps.sunflower.api.UnsplashService
+import com.google.samples.apps.sunflower.data.entity.UnsplashPhoto
 
 private const val UNSPLASH_STARTING_PAGE_INDEX = 1
