+++ app/src/main/java/com/google/samples/apps/sunflower/data/entity/UnsplashPhoto.kt
@@ -14,17 +14,19 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.entity
 
 import com.google.gson.annotations.SerializedName
 
 /**
- * Data class that represents URLs available for a Unsplash photo.
+ * Data class that represents a photo from Unsplash.
  *
- * Although several photo sizes are available, this project uses only uses the `small` sized photo.
- * For more details, consult the API documentation
- * [here](https://unsplash.com/documentation#example-image-use).
+ * Not all of the fields returned from the API are represented here; only the ones used in this
+ * project are listed below. For a full list of fields, consult the API documentation
+ * [here](https://unsplash.com/documentation#get-a-photo).
  */
-data class UnsplashPhotoUrls(
-    @field:SerializedName("small") val small: String
+data class UnsplashPhoto(
+        @field:SerializedName("id") val id: String,
+        @field:SerializedName("urls") val urls: UnsplashPhotoUrls,
+        @field:SerializedName("user") val user: UnsplashUser
 )
