+++ app/src/main/java/com/google/samples/apps/sunflower/data/entity/UnsplashPhotoUrls.kt
@@ -14,19 +14,17 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.entity
 
 import com.google.gson.annotations.SerializedName
 
 /**
- * Data class that represents a photo from Unsplash.
+ * Data class that represents URLs available for a Unsplash photo.
  *
- * Not all of the fields returned from the API are represented here; only the ones used in this
- * project are listed below. For a full list of fields, consult the API documentation
- * [here](https://unsplash.com/documentation#get-a-photo).
+ * Although several photo sizes are available, this project uses only uses the `small` sized photo.
+ * For more details, consult the API documentation
+ * [here](https://unsplash.com/documentation#example-image-use).
  */
-data class UnsplashPhoto(
-    @field:SerializedName("id") val id: String,
-    @field:SerializedName("urls") val urls: UnsplashPhotoUrls,
-    @field:SerializedName("user") val user: UnsplashUser
+data class UnsplashPhotoUrls(
+    @field:SerializedName("small") val small: String
 )
