+++ app/src/main/java/com/google/samples/apps/sunflower/data/entity/UnsplashUser.kt
@@ -14,18 +14,23 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.entity
 
 import com.google.gson.annotations.SerializedName
 
 /**
- * Data class that represents a photo search response from Unsplash.
+ * Data class that represents a user from Unsplash.
  *
  * Not all of the fields returned from the API are represented here; only the ones used in this
  * project are listed below. For a full list of fields, consult the API documentation
- * [here](https://unsplash.com/documentation#search-photos).
+ * [here](https://unsplash.com/documentation#get-a-users-public-profile).
  */
-data class UnsplashSearchResponse(
-    @field:SerializedName("results") val results: List<UnsplashPhoto>,
-    @field:SerializedName("total_pages") val totalPages: Int
-)
+data class UnsplashUser(
+    @field:SerializedName("name") val name: String,
+    @field:SerializedName("username") val username: String
+) {
+    val attributionUrl: String
+        get() {
+            return "https://unsplash.com/$username?utm_source=sunflower&utm_medium=referral"
+        }
+}
