+++ app/src/main/java/com/google/samples/apps/sunflower/data/repository/PlantRepository.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.repository
 
+import com.google.samples.apps.sunflower.data.PlantDao
 import javax.inject.Inject
 
 /**
@@ -38,8 +39,10 @@ class PlantRepository @Inject constructor(
         @Volatile private var instance: PlantRepository? = null
 
         fun getInstance(plantDao: PlantDao) =
-                instance ?: synchronized(this) {
-                    instance ?: PlantRepository(plantDao).also { instance = it }
+                instance
+                        ?: synchronized(this) {
+                    instance
+                            ?: PlantRepository(plantDao).also { instance = it }
                 }
     }
 }
