+++ app/src/main/java/com/google/samples/apps/sunflower/viewmodels/GardenPlantingListViewModel.kt
@@ -16,12 +16,13 @@
 
 package com.google.samples.apps.sunflower.viewmodels
 
+import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import com.google.samples.apps.sunflower.data.GardenPlantingRepository
-import com.google.samples.apps.sunflower.data.PlantAndGardenPlantings
+import com.google.samples.apps.sunflower.data.repository.GardenPlantingRepository
+import com.google.samples.apps.sunflower.data.entity.PlantAndGardenPlantings
 
-class GardenPlantingListViewModel internal constructor(
+class GardenPlantingListViewModel @ViewModelInject constructor(
     gardenPlantingRepository: GardenPlantingRepository
 ) : ViewModel() {
     val plantAndGardenPlantings: LiveData<List<PlantAndGardenPlantings>> =
