+++ app/src/main/res/layout/list_item_plant.xml
@@ -24,7 +24,7 @@
             type="android.view.View.OnClickListener"/>
         <variable
             name="plant"
-            type="com.google.samples.apps.sunflower.data.Plant"/>
+            type="com.google.samples.apps.sunflower.data.entity.Plant"/>
     </data>
 
     <com.google.samples.apps.sunflower.views.MaskedCardView
