+++ app/src/test/java/com/google/samples/apps/sunflower/data/GardenPlantingTest.kt
@@ -16,6 +16,7 @@
 
 package com.google.samples.apps.sunflower.data
 
+import com.google.samples.apps.sunflower.data.entity.GardenPlanting
 import org.hamcrest.CoreMatchers.equalTo
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertThat
